backend-python/requirements.txt
# Flask Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# AWS SDK
boto3==1.34.0

# Environment variables
python-dotenv==1.0.0

# Input validation
marshmallow==3.20.1

# Rate limiting
Flask-Limiter==3.5.0

# Development server
gunicorn==21.2.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0

# Utilities
python-dateutil==2.8.2

